prplmesh-hostapd: added vap_indication to load()
originally bss= was the only sign for
vap configuration inside the hostapd file.
this was based on intel's hostapd version.
the hostapd upstream indicates vap configuration
with interface= sign.
load() now seeks for the vap sign as set
by the user.
also in this commit:
- removed namespace config from prplmesh::hostapd.
left with only the two and class Configuration.
- clang format fixes
https://jira.prplfoundation.org/browse/PPM-249

Signed-off-by: Ran Regev <ran.regev@devalore.com>