agent: agent_ucc_listener: handle the backhaul sta mac
When we get dev_get_param with the "macaddr" parameter and no SSID, we
have to reply with the MAC address of the backhaul sta, and not the
MAC address of the radio.

From the CAPI specification:

When SSID is not specified for the parameter "macaddr", it implies
that device should return backhaul STA's MAC address.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>