bcl: remove timeout handling from event loop
https://jira.prplfoundation.org/browse/PPM-18

This is a preparatory commit.

As discussed in code reviews, associating a timeout to a read or write
operation on a specific file descriptor is doubtedly useful and adds
unnecessary complexity to the reactor design pattern.

Moreover, current implementation of this feature has some flaws. This
is not yet a problem because timeouts are not being used anywhere.

As discussed in architecture meetings related to socket refactorings,
operation timeouts and periodic actions will be handled by a new yet
to be implemented Timer class (a wrapper around timerfd_create).
Therefore current per-file-descriptor timeout handling will be removed.

In order to use the current event loop with the new sockets API to
implement the interface state monitor, some modifications are required
in the event loop implementation. To simplify this task and avoid
having to fix code that will be removed after all, this preparatory
commit removes the timeout handling mechanism.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>