transport: use new InterfaceStateManager
https://jira.prplfoundation.org/browse/PPM-18

Use the InterfaceStateManager class in the ieee1905_transport process
as a new mechanism to read and detect changes in the state of the
network interfaces. Existing code in `ieee1905_transport_netlink.cpp`
is deprecated and will be removed in next commit.

InterfaceStateManager uses interface names. Since we should use
interface names everywhere rather than interface indexes (because
interfaces are not renamed while prplMesh is running and conversely,
indexes can change, and because names are easier to debug than
indexes), modify existing code to favor names over indexes. BTW, this
change also fixes a bug in calls to `handle_interface_status_change`
that use a file descriptor when they shall be using an interface index.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>