ci: bump the easymesh_cert version
sniffer: use the tshark version from the environment.

See the following merge request to have all the details about the
issue:
https://git.prpl.dev/prplmesh/wfa-certification/sniffer_ucc_agent/-/merge_requests/1

In a nutshell: the tshark version bundled in the certification
repository has issues parsing some packets. Since we need to switch a
new major version, it's easier to just rely on the version present on
the environment instead of updating the bundled one.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>