agent: disconnect wireless hal on dev_reset
Previously the wireless hal was only disconnected in dev_reset if the
FSM state was already is something that comes after ENABLE.

There are some cases however where the disconnect didn't happen.

For example, if the connection is active and we stop and restart
prplMesh, the next dev_reset will not disconnect the wireless
link.

When we are in dev_reset, disconnect the wireless hal if it's active,
regardless of the FSM state.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>