openwrt: nl80211: add a 'debug' config
Some tools like wpa_cli are not strictly required by prplMesh, so they
are not dependencies of the openwrt package itself. They can be useful
to have for development though, so add them here.

Also include the optional packages we add to 'prplwrt-version', to
make sure prplwrt is flashed again when we add or remove a package.

While we're at it remove base_default_config, since it's unused.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>