son_master: activate reconnected agent on intel-slave-join
This change is to resolve an issue where the controller treats a slave
as not-active even after slave-join completes.
In the case where wireless-BH is lost, on the next topology response
received for one of the slaves, the other slave that is not part of the
topology-response is set to not-active in the controller nodes-map.
That slave is later reported as part of one of the following
topology-responses - but the only flow that enables the slaves is from
the slave join.

For non-intel-slave-join the slave is set to active in the node-map on
each join, regardless of its previous value.

For intel-slave-join there are 2 flows:
1. slave join is handled as part of fool son-slave FSM flow (on init or
due to slave-reset).
2. slave join request is sent with a flag "reconf" which means the
slave is already operational and only re-establishes a connection to
the controller.

In case 1, the channel-selection-task is triggered which sets the
slave's active flag to true as part of the task.
In case 2, the channel-selection-task is not triggered (which is
expected since the hostapd is already up) - but no other flow is
setting the slave to active in the nodes-map.

The side effects of that are:
- Any operation that verifies the slave is active is suppressed.
- The AP is not showing in bml-conn-map.

The solution:
For intel-slave-join, activate the hostap (in node map) at the end of
slave-join if the channel-selection-task is not triggered (which means
if the slave-reconf flag is set).
(For non-intel-slave-join no change is required)

Signed-off-by: Adam Dov <adamx.dov@intel.com>