framework: split the bpl components
The current way the bpl folder is organized doesn't make it easy to
choose between different implementation of its component.

For example, some platforms using packet accelerator may not be able
to use the arp monitor at all, while we would still like them to use
the uci implementation of the 'cfg' component.

Splitting first by component and then by implementation makes it more
flexible. We can for example choose to use the uci version of 'cfg',
while still keeping the linux version of 'arp'.

In addition to moving folders, the following changes are also
introduced in this commit:

  - The main "bpl.cpp" is also moved to "cfg" (while previously it was
    under uci/).
  - The CMakeLists.txt file is adapted to use the new folder structure.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>