bcl: return address length in receive_from()
A Unix socket abstract address is a string that is not 0-terminated.
Similar for an unnamed address. So we need to return address length as
well - for example by storing it in the address structure.

Socket::Address has been modified to allow for setting the length of
the address. A new method to get the size of the address has been
created (size cannot be changed).

Method SocketConnectionImpl::receive_from() has been modified to set
the length of the address.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>