client steering task: set friendliness according to steer result
To better determine the client's remaining timelife, the friendliness
status is used. If a client is unfriendly it's max timelife will be one
day (86400 seconds), or how ever is defined in the configuration.
Friendliness is determined according to how a client reacts to the
steering request. If the steering fails the client will count as
unfriendly.

Set friendliness status according to how the clients steers

Signed-off-by: itay elenzweig <itayx.elenzweig@intel.com>