controller: master: replace unconfigured-parameter occurances
The extension of eClientSelectedBands with the
addition of Unknown option band.
The option is supposed to take an unknown client
and assume it may have indefinite limit of bands option.

Apply eClientSelectedBands::SelectedBand_Unknown over the
db, the son_management, the optimal_path task and the
node files in relevant places.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>