gitlab-ci: run expensive jobs only manually
We are running a lot of pipelines and our build servers are getting
overloaded. To reduce the load, run the expensive jobs only manually, or
when it is part of a merge train. In other words, the full tests are
only done when we're about to merge - until then, tests can be triggered
manually to be sure the merge will indeed be possible.

Note that this relies on CI_MERGE_REQUEST_TARGET_BRANCH_SHA to be set
only for merge trains ("merge result pipelines"). This may change in the
future - in that case, we'll need to revisit these conditions.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>