gitlab-ci: run expensive jobs only manually
We are running a lot of pipelines and our build servers are getting
overloaded. To reduce the load, run the expensive jobs only manually, or
when the "ready for merge" label is set on the merge request.

The "expensive" jobs at the moment are all builds except for the basic
Ubuntu 18.04 build.

The "expensive" jobs can still be triggered manually. Once the build
completes, the jobs that depend on it can also be triggered manually.
This is not ideal, but a good start.

Since with this change, only part of CI will run if the "ready for
merge" label is not set, we also add a job that verifies that the label
is indeed set. This does mean that all pipelines will be marked as
failed for merge requests that are not yet ready for merge. That's
unfortunate, but until we find a different way to block the merge in
that case, there's not much we can do about it.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>