gitlab-ci: run expensive jobs only manually
We are running a lot of pipelines and our build servers are getting
overloaded. To reduce the load, run the expensive jobs only manually, or
when the "ready for merge" label is set on the merge request.

The "expensive" jobs at the moment are all builds except for the basic
Ubuntu 18.04 build.

The "expensive" jobs can still be triggered manually. Once the build
completes, the jobs that depend on it can also be triggered manually.
This is not ideal, but a good start.

Note that since we don't set the "allow failure" option, the manual jobs
will mark the pipeline as "Blocked", which means the merge request can't
be merged until they are indeed run. That can be done manually or by
setting the "ready for merge" label.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>