docker: add rdk-b build
Add infrastructure for building RDK-B. The purpose is to be able to do
this as part of CI.

The build is done in a docker image. The image contains the
prerequisites on Ubuntu 18.04, and the RDK sources themselves. Since RDK
requires a password to access the manifest file, RDK_GIT_USER and
RDK_GIT_TOKEN environment variables are used to pass in the credentials.
A git-askpass helper script is used to give the password to git. This
could also be done with a .netrc, but git-askpass is a bit more
transparent and just as much work. For CI, we'll want to use a file for
the password instead of a plain variable, so RDK_GIT_TOKEN_FILE can be
set instead of the plain token. This can't be simply COPY'd into the
image because apparently docker always considers it "new" and will not
be able to use the cache for it. Therefore, the token file is converted
into a plaintext password outside of the container.

A build script is run inside the image, that configures for the target
device, runs bitbake to build everything, and copies out the artifacts
into a mounted volume.

To reduce the build time, we mount a cache directory into the container
which contains the downloads and sstate-cache directories that are used
by bitbake. The default is to use a directory called "rdk" at the top of
the tree, so add this to gitignore. Bitbake doesn't seem to look at the
DL_DIR and SSTATE_CACHE_DIR environment variables, so the easiest way to
link to them is to mount them in the place where they're expected.
Therefore, the directories have to be created explicitly.

Unfortunately, some of the RDK packages use an "external source" which
is in fact only meant for development. "external source" packages don't
use the sstate cache, so these packages are always rebuilt. Not much can
be done about that, unfortunately.

The same "external source" mechanism is used to build the already
checked-out source rather than the git source. For the prplmesh build,
the sstate behaviour is exactly what we want. For good measure, though,
we still remove any prplmesh artifacts from the sstate cache.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>