docker: builder/ubuntu/bionic: add runner dependencies
Add the packages used by the runner to the builder as well. This will
allow the builder to be used for running prplmesh. Note that some
packages are not actually used at the moment (e.g. netcat, ping, psmisc)
but they're convenient to have when debugging.

While we're at it, fix alphabetic orderin of libnl-genl and libnl-route.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>