agent: db: add_radio: return false on successive calls
https://jira.prplfoundation.org/browse/PPM-341

The documentation of `add_radio` states:
"If called once and then called again
the radio struct on the database will not be updated with new arguments,
and the function will return false."

But in reality it returns `true` in this case.

This commit changes the code to follow the documentation.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@gmail.com>