common: hostapd: store the 'interface=' first
When we load the hostapd configuration file, we don't currently
preserve the order of the vaps.

This is an issue in case we don't store the 'main vap' (the one using
interface=) first.

This is a temporary solution to be able to progress on wireless
backhaul.

It goes over the different vaps, store the main one first, then go
over the vaps again and store the rest.

The long term solution will be implemented as part of:
https://jira.prplfoundation.org/browse/PPM-451

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>