gitlab-ci: add RDK-B build
Add the RDK-B build to CI.

Include a download and sstate cache that is shared between different
runs by putting it in the home directory. Note that these caches are not
entirely multithread safe, but the chance of two runners creating or
accessing the same cache file at the same time is very small so let's
ignore that.

To make sure two different runs don't create two conflicting docker
images, make sure the name is unique by using the --tag option.

Note that the docker image is intentionally not created in the "images"
pipeline and is not pushed to the registry, because it's not entirely
clear if RDK considers this public (even though all source repositories
are indeed public, the manifest file is not).

This build is considered "expensive" and is only run if the ready for
merge tag is set.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>