ieee1905_transport: network: split handle_interface_status_change
handle_interface_status_change handles both the activate and deactivate
case, but they use completely separate code paths. Therefore, split it
up into separate activate and deactivate functions.

This allows update_network_interfaces to reuse the activate_interface
function instead of copying it.

While we're at it, modify some of the logging to report the interface
name in addition to the file descriptor.

As a side effect, the on_error handler no longer captures the ifname
temporary, but instead captures the interface object which is guaranteed
to have a longer lifetime than the handler. This fixes
https://jira.prplfoundation.org/browse/PPM-426

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>