docker: make builder image usable as runner
Add the entrypoint.sh from the runner image as a script
start-prplmesh.sh that is installed in the builder image. Merge the
start-agent and start-agent-controller scripts into it by directly
calling prplmesh_utils.sh with the appropriate --mode argument. Since
this script uses bash anyway, we can use bash arrays to make sure of
proper quoting.

Update run.sh to use the builder image instead of the runner image, but
with start-prplmesh.sh as the entrypoint.

Update the docker-compose file in a similar way.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>