gitlab-ci: remove boardfarm-test job
The boardfarm-test job has been replaced by the docker compose based
test. Therefore, there is no use to run them both.

Since boardfarm-test is actually a better name, and it was also in a
more appropriate place in the file, update that job to run dctest
instead of run_bf.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>