agent: disable the radio in order for the wps scan to finish
Disable the radio interface to make sure it's not beaconing along
while the supplicant is scanning.
Disable rest of radio interfaces to prevent stations from
connecting (there is no BH link anyway).
This is a temporary solution for axepoint (prplwrt) to pass
wbh easymesh certification 4.2.2 test.
Will be removed after PPM-643 is solved.
Also, set configuration_in_progress flag in order not to reset the
 slaves when AP_DISABLED events arrive.

Signed-off-by: Alex Kanter <alex.kanter@devalore.com>