agent: disable the radio in order for the wps scan to finish
Disable the radio interface to make sure it's not beaconing along
while the supplicant is scanning.
set configuration_in_progress flag in order not to reset the slaves.
This is a temporary solution for intel (prplwrt) drivers to pass
wbh easymesh certification 4.2.2 test.
For permanent solution need to create wps task and move all platform
specific functionality to bwl.

Signed-off-by: Alex Kanter <alex.kanter@devalore.com>