tests: always set verbose option
In the boardfarm context we always want to see the debug output, because
boardfarm takes care of capturing output properly anyway.

In CI, we also want to always see the debug output.

Only when manually running test_flows.py it makes a little bit of sense
to not show the full debug output. However, since test_flows.py is on
its way out anyway, let's ignore that.

Therefore, globally set opts.verbose to True.

The -v argument to test_flows.py is still kept for backward
compatibility, but is ignored.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>