treewide: remove USE_USER_TMP_PATH
The USE_USER_TMP_PATH CMake option sets TMP_PATH to /tmp/<user>/beerocks
instead of the normal /tmp/beerocks. The reasoning behind this is to be
able to run prplmesh on a shared server with several users. However,
this is not possible anyway: prplmesh must be run as root, and there can
be only one instance running at a time.

The USE_USER_TMP_PATH is causing a lot of issues all over. It requires
to run prplmesh as the same user as what it was built with. This, in
turn, makes it impossible to use the artifacts from CI locally.

Since this feature is anyway useless, remove it. Update all the places
that refer to /tmp/$USER/beerocks: in README.md, in dummy bwl README, in
the dummy event handling in tests, in the mapping of the log directory
volumes in docker/run.sh and prplmesh_compose.py, and in
prplmesh_utils.sh.

Fix some whitespace errors in top-level CMakeLists.txt (automatic by
editor).

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>