CMakeLists.txt: use origin-based rpath
We used an absolute path in the -rpath arguments. This makes the
binaries non-relocatable. Make them relocatable by using an
$ORIGIN-relative path instead. We use $ORIGIN/../lib everywhere - for
the libraries, we could have used simply $ORIGIN, but this works as well
and is simpler.

The -rpath option was set in two places: once explicitly in
CMAKE_EXE_LINKER_FLAGS, and once mediated by CMake through
CMAKE_INSTALL_RPATH. Remove the first one.

https://jira.prplfoundation.org/browse/PPM-208

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>