prplmesh_utils.sh: avoid absolute paths
The prplmesh_utils.sh script has absolute paths hardcoded in many
places, based on the CMake options. This makes it impossible to move the
prplmesh tree to a different location, e.g. to use the CI results in a
different context.

Make all paths in prplmesh_utils.sh relative. Replace PRPLMESH_BIN_DIR
with PRPLMESH_DIR and use that everywhere, to avoid using .. in any of
the paths. This looks a bit cleaner.

https://jira.prplfoundation.org/browse/PPM-208

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: dfons25 <davidmgafonso@gmail.com>