docker: builder/ubuntu/bionic: split packages in different types
With the merged builder and runner docker images, a lot of packages need
to be installed and it's hard to know why they are needed. To make this
more understandable, split the list of packages into packages needed for
building prplMesh, for building Ambiorix, for test (i.e. running unit
tests), for run (i.e. running prplMesh), and for debug (additional tools
that are useful for debugging).

Since Dockerfile doesn't have the concept of variables, abuse the ARG
command to make this split.

While we're at it, remove three packages that aren't needed for
anything: autoconf, autogen, and libtool. They used to be necessary for
building nng, but we haven't used nng since a long time.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>