CMakeLists:install tests one directory higher
The tests were installed in bin/tests. However, after switching the RUNPATH to a relative path with $ORIGIN, the executables can no longer find the shared
libraries, because they are in ../../lib, not in ../lib.

We could change the RUNPATH in the tests to ../../lib. A much simpler solution, however, is to install them one directory higher: in tests/ instead of bin/tests/Update gitlab-ci.yml accordingly.

Signed-off-by: dfons25 <davidmgafonso@gmail.com>