ci: certification: apply wireless configuration before the test
Storing the uci configuration in the prplMesh repository makes sure
the settings are always in sync with the code.

At the moment prplMesh has a couple of limitations that are taken into
account here.

For example, the name of the wireless interfaces have to be exactly
like they are here for prplMesh to work.

The plan is to first include those settings to satisfy prplMesh. Then,
every time a new issue is solved, the 4.2.2 test can be run again to
make sure nothing was broken.

The mentioned limitations are:

- The default interface names cannot be used.
  See: https://jira.prplfoundation.org/browse/PPM-594

- The UCI section names have to include "default_".
  See: https://jira.prplfoundation.org/browse/PPM-126

- The channel switching as part of 4.2.2 does not work.
  The channels are currently pinned to the channels used in 4.2.2
  until this is solved.
  See: https://jira.prplfoundation.org/browse/PPM-593

Fixes https://jira.prplfoundation.org/browse/PPM-410

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>