docker: don't use the bridge for external communication
Using the IP to the bridge means we can no longer communicate with
the container if the wired interface is removed from the said bridge.

Add the container to an additional network. It has the same name as
the original network, with "ucc" appended. This additional network
will make the container have an additional interface which we can use
for ucc communication (it won't be part of the bridge).

Regular (non-compose) Docker containers can't be started with 2
"--network" options, so the UCC network has to be added once the
container has already been created.

To avoid starting the container with only one interface and adding the
other one afterwards, first create the container, add the second
network, then start it.

'eth1' is always used for ucc communication. For non-compose
containers, it should always be the ucc network, since it's added
afterwards.

For docker-compose, the list of networks seems to be order sensitive
for recent versions of docker, so ucc should always be eth1 as well.
See the following two related links:
https://github.com/docker/compose/issues/4645
https://github.com/docker/compose/pull/5566

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>