docker: don't use the bridge for external communication
Using the IP to the bridge means we can no longer communicate with
the container if the wired interface is removed from the said
bridge (when we switch to wireless backhaul for example).

Add the containers to an additional network. It has the same name as
the original network, with "ucc" appended. This additional network
will make the container have an additional interface which we can use
for ucc communication (it won't be part of the bridge).

Regular (non-compose) Docker containers can't be started with 2
"--network" options, so the second network has to be added once the
container has already been created. Docker compose containers can be
assigned multiple networks directly.

To avoid starting the containers with only one interface and adding
the other one afterwards, first create the container, add the second
network, then start it.

Because we use the "expose" feature of docker and because the network
with the highest priority (currently always the first one) is always
the one used for external communication, we have to use 'eth0' for ucc
communication. This is why the entrypoint is updated to use eth1 for
the bridge instead of eth0, and BEEROCKS_BH_WIRE_IFACE is now set to
eth1.

For non-compose containers, it's complicated. The network interfaces
seem to be assigned based on the alphabetical order of the docker
network names. Since we rely on eth0 to be the UCC interface, we need
to make sure it's name comes comes first alphabetically. For now there
doesn't seem to be another way of making sure eth0 will be the UCC
interface before the container starts.

For docker-compose, the list of networks is order sensitive for
recent versions of docker, so the interface we use for UCC
communication will always be eth0 as long as it's the first one of the
list of networks.
See the following two related links:
https://github.com/docker/compose/issues/4645
https://github.com/docker/compose/pull/5566

Fixes https://jira.prplfoundation.org/browse/PPM-595

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>