agent: Move bridge mac initialization
On certification tests, the CAPI request from the Agent the ALID
(bridge) mac of the Agent at a very early stage, before the link to
the controller is being established. But currently, the bridge mac
is being updated on the AgentDB after the backhaul link is established.
Since this PR will remove the from the agent_ucc_listener the local
bridge mac class member and its manual initialization and instead
use the data from the AgentDB, need to make sure that the data
there is valid.

Add initialization of bridge mac on AgentDB, in the init state of
the son_slave.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>