bcl: move Buffer classes to a dedicated file
https://jira.prplfoundation.org/browse/PPM-591

This is a preparatory commit.

The buffer interface and implementation classes are located in sockets.h
and sockets_impl.h respectively, together with other very simple classes
related to sockets.

Since some more methods are needed in `Buffer` and `BufferImpl` (so these
classes will become more complex) and a mock implementation is going to
be required (to write unit tests for the classes that use the buffer
interface), move these classes to a dedicated file for each class.

While we are at it, add methods `length` and `shift` and create missing
unit tests for the buffer implementation class.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>