docker: compose: use templates for controller and agent
The docker-compose yaml file has an extension fields feature that makes
it possible to avoid repitition. Use this to simplify the yaml file. We
need to use a newever version of the file format to be able to do that.
The docker versions that we check for in dctest.py are high enough to
support this.

While we're at it, use the container names that are expected by
boardfarm (lan-$ID for the controller, agent-$ID for the agent).

While we're at it, expose port 8002 (the UCC port) rather than 5000 (the
beerocks_analyzer port), because we're not going to use the latter in
boardfarm tests.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>