boardfarm: use "ports" instead of "expose" for UCC port
Since we can easily request the port mapping from docker, we don't need
to "expose" the UCC port, we can use "port" instead. This makes it
easier to find the port (no need for finding the IP address). It also
makes it easier to access the port from within the boardfarm container.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>