boardfarm: use host networking
To allow boardfarm to easily access the network of the containers, give
it host networking rather than bridge networking.

Note that this doesn't mean that only a single boardfarm instance can
run on the same machine, since it can use the same techniques as
non-containerised boardfarm to retrieve the port number to connect to or
the bridge to sniff.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>