dctest: start all containers with 'docker-compose up'
When starting the agent before the controller, it takes much longer to
start up because its controller search fails. Since the order in which
containers are started by boardfarm is not really predictable, we can't
rely on starting on the correct order.

To make this easier to control, start everything together with
'docker-compose up'. Use dependencies between the startup services to
express the correct order.

Dependency is not enough. Ideally, we should use something like
wait-for-it to poll for the UCC socket to come up. However, for now,
just add a sleep to the entrypoint.

The second agent now has to be specified explicitly in the compose file.

Since docker-compose is no longer called from PrplMeshCompose, the
_docker_compose helper function can be removed.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>