bcl: create CMDU parser interface and implementation
https://jira.prplfoundation.org/browse/PPM-591

This is a preparatory commit.

Create parser for CMDU messages received through a stream-oriented
socket (which does not preserve message boundaries).

The CMDU parser will later be used whenever data is received through a
socket connection. If a valid CMDU message is contained in data
received then it will be processed according to its type.

The CMDU parser includes a `verify_cmdu` method that has been borrowed
from `socket_thread` (this whole class is deprecated and will be
eventually removed).

Signed-off-by: Mario Maz <mmazzapater@gmail.com>