agent: platform_manager: avoid class name clashing
https://jira.prplfoundation.org/browse/PPM-591

This is a preparatory commit.

Now there are two classes named `Socket`, one in bcl/network/socket.h
and the other one in bcl/network/sockets.h. The first one is deprecated
in favor of the later and will be eventually removed.

Until then and to avoid name clashing in platform_manager_thread.cpp, we
must stop using namespace `beerocks::net` and fully qualify all objects
in that namespace.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>