bcl: pass shared_ptr by value instead of by const-ref
https://jira.prplfoundation.org/browse/PPM-591

Some constructors in the sockets API accept a shared_ptr as parameter.
This shared_ptr is stored as a member variable. The parameter should be
passed in by value instead of by const-ref to make the function’s
ownership sharing explicit.

While we are at it, add some doxygen documentation to affected classes.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>