ci: change rdk-b cache directory location
Change the "build-for-rdkb" stage to use a cache path relative
to the project directory instead of the home folder of the
script executor (i.e. gitlab-runner).

This solves an issue where the runner is running inside a docker
container. Unlike the home directory, the project directory is
mounted as a host folder, which makes it available to other
containers as well (e.g. rdk-b builder docker).

Signed-off-by: Vitaly Bukhovsky <vitaly@devalore.com>