treewide: Change clock type for persistent DB.
Currently the clock used for the persistent DB is the steady_clock,
This means that the clock is zeroed to the system's start time (it's
epoch is set to when the system starts during boot).
This is an issue as the clock should be retained between restarts of the
system.

- Change all persistent DB clock usage to use system_clock instead.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>