tlvf: Modify READ_ACS_REPORT_REQUEST/RESPONSE messages
Rename the "READ_ACS_REPORT_REQUEST/RESPONSE" messages to
"CHANNELS_LIST_REQUEST/RESPONSE".

Add to the response message "supported_channels" list.
This will be used on future agent channel selection task on a different PR.

The purpose of it is to create a message which includes all the supported
channels with channel rank and channel DFS state.

+ Update Auto-generated code.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>