bwl: Add channel rank to ACS report
Add reading of 'rank' field from bwl where reading ACS report and add
it to the preferred channels.
To support ranking on a channel with different bandwidths, push to the
preferred channel list channels on all bandwidths instead of 20MHz only.

Also, the "is_5ghz" flag on the radio info was calculated on every iteration
of the channel list. This does not make any sense, so move it after the
iterating on the supported channels.

Updated the print of the channel list so it will print the rank.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>