ci: support LOCAL_RUNNER_CACHE_DIR env variable
Commit 65ee9011d7c61c02212d33b5b3b12f3237be2d1b moved
the location of the RDK-B cache into the project directory.
While this resolved the issue with container-based GitLab runners,
it didn't actually act as a cache directory, as project directories
are not preserved between builds.

Solve this by introducing a new environment variable to allow
specifying a custom cache directory. When not set, the home
directory is used.

Signed-off-by: Vitaly Bukhovsky <vitaly@devalore.com>