agent: backhaul manager: unset autoconfiguration_failed before restart
`db->statuses.ap_autoconfiguration_failure` is a flag variable
that indicate that autoconfiguration has failed.
If this flag is set, backhaul manager (BHM) goes to RESTART state.

This commit unsets `ap_autoconfiguration_failure` before BHM restart,
so that the next time BHM checks for the flag it would not erroneosly
assume that autoconfigurated failed.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@gmail.com>