dctest: sniffer: Work around permission problems for sniffer files
For some reason wireshark command cannot create files inside logs
directory when it is started via subprocess.Popen() inside boardfarm
container even though the process itself is started as root. It can
create files when run from container's shell.

Change access permissions for logs directory so that wireshark can
create dump files and any cleanup job can cleanup log files created by
container processes which were run as root.

Wireshark process's error messages were not being shown when it failed.
Now they are printed as errors.

Tshark does not exit with non-zero code when it cannot find the the file
it's supposed to read. Instead it gives output as if it is an empty
.pcap file. So we have to verify that the file exists before running
tshark and raise an error if it does not exists.

Signed-off-by: Sinan Akpolat <aciliketcap@gmail.com>