agent: db: add channel switch parameters
https://jira.prplfoundation.org/browse/PPM-342

This is preparatory commit to remove `hostap_cs_params`
in son slave code.

Add following parametes to `sRadio` struct in agent db:
- channel
- bandwidth
- channel_ext_above_primary
- vht_center_frequency

The following variable has already been moved to agent db
from `hostap_params`. Not need to create a copy.
- tx_power

Following parameters are not kept anywhere.
They are sent in ACTION_CONTROL_SLAVE_JOINED_NOTIFICATION,
but the controller does not use them in this flow.
- switch_reason
- is_dfs_channel

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@gmail.com>