agent: platform_manager: replace Socket* with int
https://jira.prplfoundation.org/browse/PPM-591

This is a preparatory commit.

Currently, platform manager uses `Socket *` as a handle to accepted
socket connections. Next commits will replace this with a file
descriptor which is an `int`.

To prepare for that change, create overloads for all methods that get
executed whenever a socket event occurs (read and close) and for the
method that sends a CMDU (write). These overloaded methods use an `int`
instead of a `Socket *` so they can be used by the new sockets API.

Overloaded methods and temporary auxiliary member variables created for
this purpose will be removed at the end of PPM-591.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>