agent: Set the m_fronthaul_iface on son_slave on the constructor
Preparative commit.
On a follow-up commit, the son_slave will need to have the member "m_fronthaul_iface" to
be filled with a valid interface name before the ap_manager joins.
To do it, set the the 'm_fronthaul_iface' on son_slave on the constructor, and remove the set of it
on the join message.
Since the interface name is always set, revert the change introduced on 3d0e303478b4fb0a00d072c2023f5f6f934938d9.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>