agent: Add "zwdfs_ap()" function to ap_manager
Preparative commit.
Add "zwdfs_ap()" function to ap_manager that return if the radio is ZWDS radio or not.
This API should be called from outside, on the fronthaul_main to decide whether to
bring up the monitor thread or not.

Since this is a configuration that should not change during runtime, and to avoid making
a repeated calls to bwl, add a flag to ap_manager that holds this flag, after calling
"is_zwdfs_supported()" once on ap_manager "ATTACH" state.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>