agent: Don't bring up Monitor on ZWDFS radio
Change the logic on the fronthaul_main to not bring up the Monitor thread
when the ap_manager identify the radio as ZWDFS, since a ZWDFS interface
shall only use for ZWDFS purpose, and shall not Monitor anything by definition.
To promise that the "zwdfs_ap()" will return a valid return value, change the
ap_manager state comparison to state "OPERATIONAL" instead of  "ATTACHED".

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>