agent: Add and Set radio zwdfs on AgentDB
Add zwdfs flag on the radio struct on the AgentDB, and set it according to the
flag received on the ap_manager join message handler.

If the radio is zwdfs, then it means it should not receive messages from the
controller/broker, so add a filter to the message handlers received from the
controller that discard the message if it has been received on zwdfs radio
son_slave.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>