agent: Add "remove_radio_from_radios_list()" function to AgentDB
Preparative commit.
Add "remove_radio_from_radios_list()" function to AgentDB.
This function removes a given radio from the radio list on the database,
which is used when iterating on all radios of the agent.
It will be used on a follow-up commit to removing a zwdfs radio from the list.
Note that it does not actually remove the radio from the database, only from
a list that is used for iterations only.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>