agent: Slave moves to an operational state on zwdfs radio
Add a logic that if the son_slave is zwdfs, then after the ap_manager has joined and
the ap_manager_socket' has been initialized, the son_slave moves to the "operational"
state directly and skipping all the follow-up states.
Along with it, remove the zwdfs radio from the radios list on the AgentDB, and send
a BACKHAUL_DISABLE_RADIO message to the backhaul manager.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>