bwl: add 802.11 management frame event
Add a new BWL event for supporting propagation of 802.11 management
frames. Once a frame is received, a BWL implementation can forward it
to the AP Manager for processing.

The event contains the following information:
- Station MAC Address
- Type of the BWL Management Frame*
- Raw frame data

* BWL Management frame differ from the 802.11 management frame.
It's the responsibility of the BWL implementation to convert the raw
type into the BWL type.

Signed-off-by: Vitaly Bukhovsky <vitaly@devalore.com>