agent: Add "disabled_slave_sockets" memeber to backhaul manager
Preparative commit.
On the follow-up commit, the handler for the BACKHAUL_DISABLE_RADIO message
from the son_slave will be added.
The handler, among other things, will erase the radio from "slaves_sockets" and
"slave_iface_socket" so the zwdfs radio will not participate in any backhaul flow.
But still, we need to save a copy of the zwdfs socket, so the future channel selection
task will be able to communicate with it.
The new "disabled_slave_sockets" will be used to hold a copy of the socket that will
be deleted.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>